refactor(sync): decouple actions from old sync files, consolidate findPeriodForDate
- actions/copilot.ts: replace runCopilotSync with unified source,
  simplify triggerCopilotSync to use advisory lock from framework
- actions/anthropic-usage.ts: replace runAnthropicSync with unified
  source for syncAllAnthropicUsage; keep syncSingleUser (per-user)
- actions/invoice-sync.ts: remove duplicated findPeriodForDate
  (re-export from budget-utils.ts), use framework hash for lock ID
- budget-utils.ts: add findPeriodForDate (canonical location)
- api/invoices/ingest: import from budget-utils instead of invoice-sync

Old files kept as internal implementation:
- copilot-sync.ts: still used by github-copilot source (delegates)
- anthropic-sync.ts: syncSingleUser + anthropicToolFilter still needed
- github-sync.ts: interactive member sync UI depends on it
- invoice-sync.ts: syncInvoices with dry-run UI still needed
