+++ src/app/api/invoices/ingest/route.ts
@@ -6,7 +6,7 @@ import { invoices, billedCosts } from "@/lib/db/schema";
 import { eq } from "drizzle-orm";
 import { extractInvoiceFields } from "@/lib/invoice-extraction";
 import { getR2Client, getR2Bucket, getR2AccountId } from "@/lib/r2-client";
-import { findPeriodForDate } from "@/actions/invoice-sync";
+import { findPeriodForDate } from "@/lib/budget-utils";
 
 /** System user ID for automated/API-initiated operations */
 const SYSTEM_ADMIN_USER_ID = 1;
